+++ src/main/kotlin/sogong/sogongSpring/repository/HotPostRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.*
 
-interface HotPostRepository : JpaRepository<HotPostEntity, Int> {
+interface HotPostRepository : JpaRepository<HotPostEntity, Long> {
 }
