+++ src/main/kotlin/sogong/sogongSpring/repository/ScrapLikeRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.*
 
-interface ScrapLikeRepository : JpaRepository<ScrapLikeEntity, Int> {
+interface ScrapLikeRepository : JpaRepository<ScrapLikeEntity, Long> {
 }
